+++ ios/LinkDemo-ObjC/LinkDemo-ObjC/ViewController.m
@@ -63,6 +63,7 @@
 - (IBAction)didTapButton:(id)sender {
 #if USE_CUSTOM_CONFIG
     [self presentPlaidLinkWithCustomConfiguration];
+    // [self presentPlaidLinkWithCopyCustomization];
 #else
     [self presentPlaidLinkWithSharedConfiguration];
 #endif
@@ -111,7 +112,7 @@
     
 }
 
-#pragma mark Start Plaid Link with shared configuration from Info.plist
+#pragma mark Start Plaid Link with custom instance configuration
 - (void)presentPlaidLinkWithCustomConfiguration {
 
     // <!-- SMARTDOWN_PRESENT_CUSTOM -->
@@ -131,6 +132,60 @@
     }
     // <!-- SMARTDOWN_PRESENT_CUSTOM -->
 
+}
+
+#pragma mark Start Plaid Link with custom instance configuration including client-side customizations
+- (void)presentPlaidLinkWithCopyCustomization {
+    PLKConfiguration* linkConfiguration;
+    @try {
+        linkConfiguration = [[PLKConfiguration alloc] initWithKey:@"<#YOUR_PLAID_PUBLIC_KEY#>" env:PLKEnvironmentSandbox product:PLKProductAuth];
+
+        /*
+         NOTE: The preferred method to customize LinkKit is to use the customization feature
+         in the dashboard (http://localhost:6419/ios/dashboard-customization).
+         In the rare case where customization is necessary from within your application directly
+         and you prefer to initialize link directly using instance configuration an example
+         is given below.
+         For further details which elements can be customized on which panes please refer to the online documentation available at:
+         https://github.com/plaid/link/blob/master/ios/README.md#customization
+         */
+        // <!-- SMARTDOWN_CUSTOMIZATION -->
+        [linkConfiguration customizeWithDictionary:@{
+                                                     kPLKConnectedPaneKey: @{
+                                                             kPLKCustomizationTitleKey: @"Sign-up successful",
+                                                             kPLKCustomizationMessageKey: @"You successfully linked your account with <CLIENT>",
+                                                             kPLKCustomizationSubmitButtonKey: @"Continue"
+                                                             },
+                                                     
+                                                     kPLKReconnectedPaneKey: @{
+                                                             kPLKCustomizationTitleKey: @"Update successful",
+                                                             kPLKCustomizationMessageKey: @"You successfully updated your account credentials <CLIENT>",
+                                                             kPLKCustomizationSubmitButtonKey: @"Continue"
+                                                             },
+                                                     
+                                                     kPLKInstitutionSelectPaneKey: @{
+                                                             kPLKCustomizationTitleKey: @"Choose your bank",
+                                                             kPLKCustomizationSearchButtonKey: @"Search for your bank"
+                                                             },
+                                                     
+                                                     kPLKInstitutionSearchPaneKey: @{
+                                                             kPLKCustomizationExitButtonKey: @"Quit",
+                                                             kPLKCustomizationInitialMessageKey: @"Find your bank or credit union",
+                                                             kPLKCustomizationNoResultsMessageKey: @"Unfortunately the institution you searched for could not be found"
+                                                             },
+                                                     }];
+        // <!-- SMARTDOWN_CUSTOMIZATION -->
+
+        linkConfiguration.clientName = @"Link Demo";
+        id<PLKPlaidLinkViewDelegate> linkViewDelegate  = self;
+        PLKPlaidLinkViewController* linkViewController = [[PLKPlaidLinkViewController alloc] initWithConfiguration:linkConfiguration delegate:linkViewDelegate];
+        if (UI_USER_INTERFACE_IDIOM() == UIUserInterfaceIdiomPad) {
+            linkViewController.modalPresentationStyle = UIModalPresentationFormSheet;
+        }
+        [self presentViewController:linkViewController animated:YES completion:nil];
+    } @catch (NSException *exception) {
+        NSLog(@"Invalid configuration: %@", exception);
+    }
 }
 
 #pragma mark Start Plaid Link with an institution pre-selected